// apps/admin/src/generated/api/schemas/updateBasicSettingsDto.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Common Admin API
 * API for the common admin starter template
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBasicSettingsDtoDefaultLocale } from "./updateBasicSettingsDtoDefaultLocale";
import type { UpdateBasicSettingsDtoDefaultTheme } from "./updateBasicSettingsDtoDefaultTheme";

export interface UpdateBasicSettingsDto {
  /**
   * @minLength 1
   * @maxLength 80
   */
  siteName: string;
  /** @maxLength 160 */
  siteSubtitle: string;
  defaultLocale: UpdateBasicSettingsDtoDefaultLocale;
  defaultTheme: UpdateBasicSettingsDtoDefaultTheme;
}